Compile issues boiled down simply to me forgetting to specify some virtual fields!

+++ inc/AkariSubsystem.hpp
@@ -9,8 +9,8 @@ class AkariSubsystem {
 	public:
 		AkariSubsystem(Akari *);
 
-		virtual u16 VersionMajor() const = 0;
-		virtual u16 VersionMinor() const = 0;
+		virtual u8 VersionMajor() const = 0;
+		virtual u8 VersionMinor() const = 0;
 		// TODO: manufacturer string?
 		// TODO: product name?
